shopping-assistant/requirements.txt
langchain
langchain-openai
langchain-qdrant
langchain-community
openai
python-dotenv
qdrant-client
langgraph
pandas